ci: pin Rust 1.98 for lint jobs and fix its new clippy lints
CI's stable toolchain is 1.98, whose clippy flags lints that 1.94 does not
(72 findings). Pin the lint jobs to 1.98 so a future stable release cannot
turn main red under -D warnings without a code change; bump deliberately.

- Allow chunks_exact_to_as_chunks crate-wide (51 sites): `chunks_exact(N)`
  is equivalent, and rewriting would change slice types throughout.
- Fix the rest: checked_div for guarded u64 divisions, while-let loops,
  sort_by_key with Reverse (both sorts are stable, same order), values_mut
  and keys() for map iteration, fill(0), and redundant casts, borrows and
  into_iter.

Validated with Rust 1.98: clippy -D warnings clean on both crates, rustfmt
clean on changed files, `cargo test --release --lib --tests` passes.

+++ src/isogeny/attack.rs
@@ -140,7 +140,7 @@ pub fn rho_on_curve(curve: &SmallCurve, max_iters: u64, seed: u64) -> (Option<Rh
     };
 
     // Pick a private scalar d in [1, n−1].
-    let d = ((seed.wrapping_mul(0x9E3779B97F4A7C15) % (n - 1)) + 1) as u64;
+    let d = (seed.wrapping_mul(0x9E3779B97F4A7C15) % (n - 1)) + 1;
     let h = g.scalar_mul(&BigUint::from(d), &a_fe);
     let _ = d;
